[ADD] DLLs Folder, Build Events, Window
DLLs are no inside a specified folder that are included with the build currently for x64 debug.

It copies " pre-link " events and adds them to the root of the build, also see some additionaly changes such as>
The DLLs is what the project looks for first rather than the default .exe

Build Events for Assets folder
Now includes the folder, sub-folders and its assets directly to the debug build.

Window is now a specified class instead of scattered in the main file.

Window Files needs improvements to Initialization and also rescaling the Application.