Control Move in EventHandler
Currently the Movement is handled within the eventhandler, though it sets the values and information is retrieved separately after events are processed  and by using the " get value " prompt.

This should be changed when player is fully developed as a separate class instead.

Also fixed  InputSystem to better use  Pressed and released.